Fix integer overflow bugs affecting large ISO files - could lead to errors reading meta-data or reading garbage file contents.

+++ src/Iso9660/CDReader.cs
@@ -99,7 +99,7 @@ namespace DiscUtils.Iso9660
             }
 
             // Skip to Path Table
-            data.Position = _volDesc.LogicalBlockSize * _volDesc.TypeLPathTableLocation;
+            data.Position = _volDesc.LogicalBlockSize * (long)_volDesc.TypeLPathTableLocation;
             byte[] pathTableBuffer = new byte[_volDesc.PathTableSize];
             data.Read(pathTableBuffer, 0, pathTableBuffer.Length);
